# server/openapi3_0.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TODO API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:6969",
            "description": "This is our local dev environment"
        }
    ],
    "paths": {
        "/todolist": {
            "get": {
                "summary": "Get all TODO lists",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "Buy groceries"
                                    },
                                    {
                                        "id": 2,
                                        "name": "Finish coding"
                                    }
                                ]
                            }
                        }
                    }
                },
                "tags": [
                    "Todolist"
                ]
            },
            "post": {
                "summary": "Create a new Todolist",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "todo"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful creation",
                        "content": {
                            "application/json": {
                                "example": {
                                    "name": "Monday"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Todolist"
                ]
            }
        },
        "/todolist/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "The Id of the Todolist",
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "delete": {
                "summary": "Delete a Todolist by ID",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": {
                                    "id": 1,
                                    "name": "Phad thai"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "TODOlist does not exist",
                        "content": {
                            "application/json": {
                                "example": "Todolist not found"
                            }
                        }
                    }
                },
                "tags": [
                    "Todolist"
                ]
            }
        },
        "/linkedin/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "The Id of the Todolist",
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "get": {
                "summary": "Get all TODOs from specific list",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "list_id": 2,
                                        "list_name": "Tis",
                                        "todo": {
                                            "todo": "Run",
                                            "time": 60
                                        }
                                    },
                                    {
                                        "list_id": 2,
                                        "list_name": "Tis",
                                        "todo": {
                                            "todo": "Run",
                                            "time": 60
                                        }
                                    }
                                ]
                            }
                        }
                    }
                },
                "tags": [
                    "Linkedin"
                ]
            }
        },
        "/todo": {
            "get": {
                "summary": "Get all TODOs",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": [
                                    {
                                        "id": 1,
                                        "todo": "Buy groceries",
                                        "time": 30,
                                        "isDone": false
                                    },
                                    {
                                        "id": 1,
                                        "todo": "Buy groceries",
                                        "time": 30,
                                        "isDone": false
                                    }
                                ]
                            }
                        }
                    }
                },
                "tags": [
                    "TODOs"
                ]
            }
        },
        "/todos": {
            "post": {
                "summary": "Create a new TODO",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "list_id": {
                                        "type": "integer",
                                        "description": "The ID of the TODO list"
                                    },
                                    "todo": {
                                        "type": "object",
                                        "properties": {
                                            "todo": {
                                                "type": "string",
                                                "description": "The TODO text"
                                            },
                                            "time": {
                                                "type": "integer",
                                                "description": "Estimated time in minutes"
                                            },
                                            "isDone": {
                                                "type": "boolean",
                                                "description": "If the task is done"
                                            }
                                        },
                                        "description": "The TODO details"
                                    }
                                },
                                "required": [
                                    "todo"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful creation",
                        "content": {
                            "application/json": {
                                "example": {
                                    "id": 3,
                                    "listid": 8,
                                    "todo": {
                                        "todo": "Read a book",
                                        "time": 60
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "List not found",
                        "content": {
                            "application/json": {
                                "example": "List with id 404 does not exist"
                            }
                        }
                    }
                },
                "tags": [
                    "TODOs"
                ]
            }
        },
        "/todos/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "The ID of the TODO item",
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "get": {
                "summary": "Get a TODO by ID",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": {
                                    "id": 1,
                                    "todo": "Buy groceries",
                                    "time": 30
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "TODOs"
                ]
            },
            "delete": {
                "summary": "Delete a TODO by ID",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": {
                                    "id": 1,
                                    "todo": "aquaman",
                                    "time": 10
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "TODO does not exist",
                        "content": {
                            "application/json": {
                                "example": "Todo not found"
                            }
                        }
                    }
                },
                "tags": [
                    "TODOs"
                ]
            },
            "put": {
                "summary": "UPDATE todo by ID",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "list_id": {
                                        "type": "integer",
                                        "description": "The ID of the list to associate with the TODO"
                                    },
                                    "todo": {
                                        "type": "object",
                                        "properties": {
                                            "todo": {
                                                "type": "string",
                                                "description": "The TODO text"
                                            },
                                            "time": {
                                                "type": "integer",
                                                "description": "Estimated time in minutes"
                                            },
                                            "isDone": {
                                                "type": "boolean",
                                                "description": "Is the task done?"
                                            }
                                        },
                                        "required": [
                                            "todo"
                                        ]
                                    }
                                },
                                "required": [
                                    "list_id",
                                    "todo"
                                ],
                                "example": {
                                    "list_id": 1,
                                    "todo": {
                                        "todo": "Updated Task",
                                        "time": 45
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "example": {
                                    "id": 1,
                                    "list_id": 1,
                                    "todo": {
                                        "todo": "Updated Task",
                                        "time": 45
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "TODO does not exist",
                        "content": {
                            "application/json": {
                                "example": "Todo not found"
                            }
                        }
                    }
                },
                "tags": [
                    "TODOs"
                ]
            }
        },
        "/todoStatus/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "The ID of the TODO item",
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "put": {
                "summary": "Update status",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "isDone": {
                                        "type": "boolean"
                                    }
                                },
                                "required": [
                                    "isDone"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful creation",
                        "content": {
                            "application/json": {
                                "example": {
                                    "listid": 8,
                                    "todo": {
                                        "isDone": "true"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "List not found",
                        "content": {
                            "application/json": {
                                "example": "List with id 404 does not exist"
                            }
                        }
                    }
                },
                "tags": [
                    "TODOs"
                ]
            }
        }
    },
    "tags": [
        {
            "name": "TODOs",
            "description": "Operations related to TODOs"
        },
        {
            "name": "Todolist",
            "description": "Operations related to Todolist"
        },
        {
            "name": "Linkedin",
            "description": "Operations related to Linkedin"
        }
    ]
}